util Path improvements
generic path definitions improved. Now, RPath (was TPath) basically
acts like a reg-exp + operator, that is, takes a relation R and defines
a list 1 or more objects long where the relation applies to each i'th
and i+1'th (i.e. list a, R a[i] a[i+1]), as well as reflexively if the
relation R is reflexive. Also, now the RPath'', or the RPath that
includes a specification of the head and tail element, can be applied
to a list 1 long to account for a possible reflexive example/instance,
now that the physical structure of the list is not specified, but only
what the head and tail elements equal instead. (For example, previously
the list [x] would not be valid for RPath'' R [x] x x, even if R was
reflexive. now it can)