Almost finished finite_set_list_length
Nearing completion of the theorem that is the super high level support
which should allow the infamous minimal member theorem to be proved. It
just says, given a list with no duplicates, if we're building that list
from a set, the list can be no longer than the size of the set. Almost
there. It's proven given one assumption/supporting lemma I used "cut"
to assume for a bit.